+++ R/dplyr-product.R
@@ -27,6 +27,10 @@
 #' orders |> run() |> collect()
 NULL
 
+#' @rdname product-dplyr
+#' @export
+dplyr::filter
+
 dplyr_product_step <- function(x, verb, args) {
   add_transform(
     x,
@@ -41,7 +45,7 @@ mutate.tw_product <- function(.data, ...) {
   dplyr_product_step(.data, "mutate", rlang::enquos(...))
 }
 #' @rdname product-dplyr
-#' @exportS3Method dplyr::filter
+#' @export
 filter.tw_product <- function(.data, ..., .by = NULL, .preserve = FALSE) {
   args <- rlang::enquos(...)
   if (!missing(.by)) {